Database: second pass at file reorganization.
* Update more namespaces & use statements
* Update more extends
