+++ includes/database/rows/class-order-item.php
@@ -3,12 +3,12 @@
  * Order Item Database Object Class.
  *
  * @package     EDD
- * @subpackage  Database\Objects
+ * @subpackage  Database\Rows
  * @copyright   Copyright (c) 2018, Easy Digital Downloads, LLC
  * @license     http://opensource.org/licenses/gpl-2.0.php GNU Public License
  * @since       3.0
  */
-namespace EDD\Database\Objects;
+namespace EDD\Database\Rows;
 
 // Exit if accessed directly
 defined( 'ABSPATH' ) || exit;
